requirements.txt
python-dotenv==0.21.0
psycopg2==2.9.5
django==4.1.3
djangorestframework==3.14.0
django-cors-headers==3.13.0
pandas==1.5.1
peakdetect==1.2.0